% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFromDist.R
\name{graphFromDist}
\alias{graphFromDist}
\title{Return a graph with given size and edge distribution
Is an inputGenerator for Time Machine}
\usage{
graphFromDist(size, distrfit)
}
\arguments{
\item{size}{- how many vertices should the graph have?}

\item{distrfit}{- return value of fitdistrplu::fitdistr call. This function only requires fields "distname" and "estimate"}
}
\description{
Return a graph with given size and edge distribution
Is an inputGenerator for Time Machine
}
