% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGenerator.R
\name{graphGenerator}
\alias{graphGenerator}
\title{Return a graph with given size and edge distribution.
Is inputGenerator when provided with distrSampler}
\usage{
graphGenerator(size, params)
}
\arguments{
\item{size}{- how many vertices should the graph have?}

\item{params}{- should contain a function distrSampler(size) returning vector of size samples from the desired
degree distribution}
}
\description{
Return a graph with given size and edge distribution.
Is inputGenerator when provided with distrSampler
}
