% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPOETSGraph.R
\name{readPOETSGraph}
\alias{readPOETSGraph}
\title{given a path to file with columns "from" and "to", returns igraph object}
\usage{
readPOETSGraph(path)
}
\description{
given a path to file with columns "from" and "to", returns igraph object
}
